created custom hook for playerpage
