+++ Components/PlayerStats/index.js
@@ -1,22 +1,22 @@
 import { StyledSelect, NoData } from "@/styles";
 import { SelectionContainer, StatsContainer } from "./Styling";
 import { StatsListItem } from "./StatsList";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { Button } from "@mui/material";
+import { usePlayerStatsData } from "@/utilities/hooks/PlayerPage/fetchplayerdata.js";
 
 export default function PlayerStats({ id, CURRENT_SEASON }) {
   const [season, setSeason] = useState(CURRENT_SEASON);
-  const [playerStats, setPlayerStats] = useState([]);
-  const [teamNames, setTeamNames] = useState({});
-  const [showNoData, setShowNoData] = useState(false);
-  const statsURL = "https://www.balldontlie.io/api/v1/stats";
-  const teamNamesURL =
-    "https://www.balldontlie.io/api/v1/teams?page=1&per_page=100";
+  const { playerStats, teamNames, showNoData, fetchStats } = usePlayerStatsData(
+    season,
+    id
+  );
 
   let seasons = [];
   for (let year = CURRENT_SEASON; year >= 1946; year--) {
     seasons.push(year);
   }
+
   const selectYears = seasons.map((season) => ({
     value: season,
     label: season.toString(),
@@ -25,32 +25,6 @@ export default function PlayerStats({ id, CURRENT_SEASON }) {
   function handleSeasonChange(selectedOption) {
     setSeason(selectedOption.value);
   }
-
-  async function fetchStats() {
-    const response = await fetch(
-      `${statsURL}?seasons[]=${season}&player_ids[]=${id}&per_page=100`
-    );
-    const { data } = await response.json();
-    setPlayerStats(data);
-    setShowNoData(data.length === 0);
-  }
-
-  async function fetchTeamNames() {
-    const response = await fetch(teamNamesURL);
-    const { data } = await response.json();
-    const names = data.reduce(
-      (acc, team) => ({
-        ...acc,
-        [team.id]: team.full_name,
-      }),
-      {}
-    );
-    setTeamNames(names);
-  }
-
-  useEffect(() => {
-    fetchTeamNames();
-  }, []);
 
   return (
     <>
